Bugs/bad withdraw behavior (#1199)
The goal of this PR is to fix a bug where CapacityDetails.remaining
capacity was never being decreased on a withdraw.

Fixes broken mocks because of the rebase (sorry), and updates
Capacity Design Doc to reflect the CapacityDetails new names and fields
(also sorry, I missed it in the last rename somehow).

This affected a number of tests that assumed the behavior was correct.